Enable text insertion

+++ src/sc_parse.h
@@ -63,7 +63,7 @@ extern SCBlock *find_last_child(SCBlock *bl);
 
 extern void sc_block_set_options(SCBlock *bl, char *opt);
 extern void sc_block_set_contents(SCBlock *bl, char *con);
-extern void sc_insert_text(SCBlock *b1, int o1, const char *t);
+extern void sc_insert_text(SCBlock *b1, size_t o1, const char *t);
 extern void sc_insert_block(SCBlock *b1, int o1, SCBlock *ins);
 extern void sc_delete_text(SCBlock *b1, int o1, SCBlock *b2, int o2);
